Made minor refactoring
Reordered import statements. Changed package structure.
