+++ bot/src/main/java/edu/scrapper/bot/BotApplication.java
