+++ bot/src/main/java/edu/scrapper/bot/configuration/ApplicationConfig.java
