+++ scrapper/src/main/java/edu/scrapper/ScrapperApplication.java
@@ -1,4 +1,4 @@
-package edu.java;
+package edu.scrapper;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
