+++ scrapper/src/main/java/edu/scrapper/configuration/ApplicationConfig.java
@@ -1,8 +1,5 @@
-package edu.java.configuration;
+package edu.scrapper.configuration;
 
-import edu.java.service.client.GitHubClient;
-import org.springframework.boot.ApplicationRunner;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
