+++ scrapper/src/main/java/edu/scrapper/configuration/ClientConfig.java
@@ -1,6 +1,6 @@
-package edu.java.configuration;
+package edu.scrapper.configuration;
 
-import edu.java.service.client.GitHubClient;
+import edu.scrapper.service.client.GitHubClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.reactive.function.client.WebClient;
