+++ scrapper/src/main/java/edu/scrapper/dto/github/PullRequestCommentTo.java
@@ -1,4 +1,4 @@
-package edu.java.dto.github;
+package edu.scrapper.dto.github;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
