+++ scrapper/src/main/java/edu/scrapper/scheduling/LinkUpdateScheduler.java
@@ -1,10 +1,10 @@
-package edu.java.scheduling;
+package edu.scrapper.scheduling;
 
+import java.util.concurrent.TimeUnit;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import java.util.concurrent.TimeUnit;
 
 @Component
 public class LinkUpdateScheduler {
